Removes the verbatim copy of the grammar in text.md. (#110)
There are two problems with copying the grammar for display on text.md:
1. When changes are made to the grammar, we have to remember to change text.md.
2. The copy can't actually be made verbatim, because this leads to rendering problems (see #103). So we'd have to remember not only to copy the new grammar to text.md, but also to fix the rendering (assuming it's fixable).

Since the authoritative source of the text grammar is only one click away, I think the best solution is to simply remove the copy. If it were possible to transclude from the grammar page that would be another option, but I haven't found a good way to do this.