Amend Supported Unicode Text Encodings
This change removes the stated support for UTF-16 and UTF-32,
reflecting a similar change in the JSON specification over time.

RFC 7159 Sec. 8.1 specified that JSON text SHALL be encoded in UTF-8,
UTF-16, or UTF-32. In 2017 RFC 8259 revised Sec. 8.1 to reflect that
JSON text MUST be encoded as UTF-8 when "exchanged between systems
that are not part of a closed ecosystem"

No Ion implementations support non-UTF-8 encodings.