Adds new diagrams for timestamps
Makes the diagrams vertical to make the endian aspects easier to read.
Also changes the text to emphasize that the bit-packed parts of
timestamp are interpreted within a `FixedUInt`.  Updates the text around
long-form timestamp to use *scale* instead of *exponent* to be more
consistent with nomenclature in the wild.

Decided against the use of *mantissa* to describe fractional seconds,
because while the former is concise, it doesn't really reduce text and
may be harder for readers to understand.

Fixes #220, Fixes #221