Add rough "continuation trace" to bottom values.
This is intended to help pinpoint where runtime errors occur in the
model.

It's not terribly satisfying at the moment, since it only marks the
outermost expression of user-defined functions, not any inner expression
layers. Still, perhaps better than nothing; at least we know the
function where the problem occurred.

This might work out better if we had #%app so we could rewrite each
application to record the frame, or if we could capture the Fusion stack
trace natively.